refactor: elapsedTime 함수 utils로 분리
